chore(test): reestructure test files by module and update Jest config
