+++ __mocks__/react-to-pdf.ts
@@ -1,0 +1,3 @@
+// This file is a mock for the react-to-pdf library.
+
+export default {};
