+++ jest.config.js
@@ -5,6 +5,10 @@ export default {
     "^.+.tsx?$": ["ts-jest", {}],
   },
   moduleNameMapper: {
-    "\\.(css)$": "<rootDir>/src/__mocks__/cssMock.ts",
+    "\\.(css)$": "<rootDir>/__mocks__/cssMock.ts",
   },
+  testMatch: [
+    "**/__tests__/**/*.test.[jt]s?(x)",
+    "**/?(*.)+(spec|test).[tj]s?(x)",
+  ],
 };
