--- src/__mocks__/react-to-pdf.ts
@@ -1,1 +1,0 @@
-export default {};
