+++ tsconfig.json
@@ -10,8 +10,11 @@
     "moduleResolution": "node",
     "resolveJsonModule": true,
     "isolatedModules": true,
-    "noEmit": true
+    "noEmit": true,
+    "paths": {
+      "@Types/*": ["./@types/*"]
+    }
   },
-  "include": ["src"],
+  "include": ["src", "@types", "__tests__", "__mocks__"],
   "exclude": ["node_modules", "dist"]
 }
